requirements.txt
crewai[tools]
pdf2image
pdfplumber
pi-heif
pillow
pymupdf
pymupdf4llm
pypdf
redis
unstructured
unstructured-inference
unstructured-pytesseract
qdrant-client
ollama
transformers
sentence-transformers
torch
fastapi
uvicorn[standard]
streamlit
requests
pyrefly
clip
openai-clip
python-dotenv
